Make circular.py parallel
